refactor: move account network row ouside of confirmations
