+++ app/components/Views/confirmations/components/modals/switch-account-type-modal/switch-account-type-modal.tsx
@@ -16,7 +16,7 @@ import { selectInternalAccounts } from '../../../../../../selectors/accountsCont
 import Spinner from '../../../../../UI/AnimatedSpinner';
 import { useStyles } from '../../../../../hooks/useStyles';
 import { useEIP7702Networks } from '../../../hooks/7702/useEIP7702Networks';
-import AccountNetworkRow from './account-network-row';
+import AccountNetworkRow from '../../../../AccountNetworkRow/account-network-row';
 import styleSheet from './switch-account-type-modal.styles';
 import Icon, {
   IconColor,
